minor cleanup pass

+++ src/index.ts
@@ -4,23 +4,24 @@ import { promises as fsa } from "node:fs";
 import edits from "./edits.js";
 import cfg from "./config.js";
 
-import "zx/globals";
+import logger from "node-color-log";
+logger.setLevel("debug");
+logger.setDate(() => new Date().toLocaleTimeString());
 
+import "zx/globals";
 $.verbose = true;
 usePowerShell();
 
-console.log("Extracting files.");
+logger.info("Extracting files.");
 
 // Cleaning up existing install/overlay directories.
-if (await fsa.stat(cfg.paths.install).catch(() => false))
-  await fsa.rmdir(cfg.paths.install, { recursive: true });
-if (await fsa.stat(cfg.paths.overlay).catch(() => false))
-  await fsa.rmdir(cfg.paths.overlay, { recursive: true });
+if (await fsa.stat(cfg.paths.install).catch(() => false)) await fsa.rmdir(cfg.paths.install, { recursive: true });
+if (await fsa.stat(cfg.paths.overlay).catch(() => false)) await fsa.rmdir(cfg.paths.overlay, { recursive: true });
 
 // Building the temp directory.
 for (const dir of Object.values(cfg.paths.temp)) {
   if (await fsa.stat(dir).catch(() => false)) {
-    console.log(`Removing existing directory: ${dir}`);
+    logger.info(`Removing existing directory: ${dir}`);
     await fsa.rmdir(dir, { recursive: true });
   }
 
@@ -33,26 +34,10 @@ for (const file of Object.keys(edits)) {
   let inputPath = path.join(wadsPath, file);
   let storePath = path.join(cfg.paths.temp.game, file);
 
-  console.log(`Extracting ${file} to ${storePath}`);
+  logger.info(`Processing file: ${file}`);
   await fsa.mkdir(storePath, { recursive: true });
 
-  await $`${cfg.paths.wad2bin} ${inputPath} ${storePath}`;
-}
-
-// delete non-bin files from the temp game directory
-for (const ent of await fsa.readdir(cfg.paths.temp.game, {
-  recursive: true,
-  withFileTypes: true,
-})) {
-  if (!ent.isFile()) {
-    continue;
-  }
-
-  if (ent.name.endsWith(".bin")) {
-    continue;
-  }
-
-  await fsa.rm(`${ent.parentPath}/${ent.name}`, { force: true });
+  await $`${cfg.paths.wadTool} extract --input ${inputPath} --output ${storePath} --filter-type bin`;
 }
 
 await $`${cfg.paths.ritoBin} -i bin -o json -r ${cfg.paths.temp.game} ${cfg.paths.temp.text}`;
@@ -63,17 +48,17 @@ let tempMods = cfg.paths.temp.mods;
 await fsa.mkdir(tempMods, { recursive: true });
 
 for (const [wadPath, bins] of Object.entries(edits)) {
-  console.log("Processing edits for:", wadPath);
+  logger.info(`Processing edits for: ${wadPath}`);
 
-  for (const bin of Object.keys(bins)) {
+  for (const [bin, edits] of Object.entries(bins)) {
     let ent = path.join(tempText, wadPath, bin);
-    console.log(`Editing: ${ent}`);
+    logger.info(`Editing file: ${ent}`);
 
     let data = await fsa.readFile(ent);
     let json = JSON.parse(data.toString());
     await fsa.writeFile(ent, JSON.stringify(json, null, 2));
 
-    for (const edit of edits[wadPath][bin]) {
+    for (const edit of edits) {
       edit(json.entries);
     }
 
@@ -81,14 +66,14 @@ for (const [wadPath, bins] of Object.entries(edits)) {
 
     let root = path.join(tempMods, wadName);
     let full = path.join(tempMods, wadName, bin);
-    console.log(`Writing: ${full}`);
+    logger.info(`Writing edited file: ${full}`);
 
     await fsa.mkdir(root, { recursive: true });
     await fsa.writeFile(full, JSON.stringify(json, null, 2));
   }
 }
 
-console.log("Running ritobin...");
+logger.info("Executing ritobin.");
 await $`${cfg.paths.ritoBin} -i json -o bin -r ${cfg.paths.temp.mods} ${cfg.paths.temp.done}`;
 
 const modDir = path.join(cfg.paths.install, "mod");
@@ -102,8 +87,8 @@ await fsa.writeFile(path.join(modDir, "META", "info.json"), "");
 
 await fsa.cp(cfg.paths.temp.done, wadDir, { recursive: true });
 
-console.log("Preparing overlay.");
+logger.info("Preparing overlay.");
 await $`${cfg.paths.modTool} mkoverlay ${cfg.paths.install} ${cfg.paths.overlay} --noTFT --game:${cfg.paths.gameDir} --mods:mod`;
 
-console.log("Initializing overlay.");
+logger.info("Initializing overlay.");
 await $`${cfg.paths.modTool} runoverlay ${cfg.paths.overlay} . --game:${cfg.paths.gameDir}`;
